backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
google-genai
python-dotenv
pydantic
fastapi-cors
